Adjusted and added needed sections for networking.

+++ src/player.rs
@@ -1,7 +1,7 @@
 use crate::light_manager::Lights;
 use crate::{
     GameState, components::Health, components::KinematicCollider, components::LightSource,
-    events::DamagePlayerEvent, player_material::PlayerBaseMaterial,
+    events::DamagePlayerEvent, player_material::PlayerBaseMaterial, net_control::NetControl, net_control::PlayerType,
 };
 use bevy::math::bounding::Aabb2d;
 use bevy::prelude::*;
@@ -68,28 +68,6 @@ impl Velocity {
     }
 }
 
-#[derive(Component)]
-pub struct NetControl {
-    player_type: PlayerControl,
-    pub net_input: u8,
-    pub player_id: u8,
-}
-impl NetControl {
-    fn new(ptype: PlayerControl, pid: u8) -> Self {
-        Self {
-            player_type: ptype,
-            net_input: 0,
-            player_id: pid,
-        }
-    }
-}
-
-#[derive(PartialEq)]
-enum PlayerControl {
-    Local,
-    Network,
-}
-
 pub fn setup_player(
     mut commands: Commands,
     mut meshes: ResMut<Assets<Mesh>>,
@@ -97,85 +75,51 @@ pub fn setup_player(
     asset_server: Res<AssetServer>,
     // query: Query<Entity, With<Camera>>,
     query: Query<Entity, With<Camera>>,
+    players: Query<Entity, With<NetControl>>,
     lights: Res<Lights>,
 ) {
     // if query.is_empty() {
     //     commands.spawn(Camera2d);
     // }
 
-    commands.spawn((
-        // For any entities that we want to have lighting,
-        // add the following two components.
-        Mesh2d(meshes.add(Rectangle::default())),
-        MeshMaterial2d(materials.add(PlayerBaseMaterial {
-            // Generally, only change what's inside the 'lighting' struct and the 'texture' and 'normal' parameters.
-            color: LinearRgba::BLUE,
-            texture: Some(asset_server.load("player/player_albedo.png")),
-            lighting: crate::player_material::Lighting {
-                // 'ambient_reflection_coefficient' and 'ambient_light_intensity' do the same thing.
-                // Should be 0 for everything except the player.
-                // 'diffuse_reflection_coefficient' is how much not-shiny light is reflected back.
-                // 'shininess' is what it sounds like. Higher number = shinier.
-                ambient_reflection_coefficient: 0.1,
-                ambient_light_intensity: 0.1,
-                diffuse_reflection_coefficient: 1.0,
-                shininess: 40.0,
+    for i in players {
+        commands.entity(i).insert((
+            // For any entities that we want to have lighting,
+            // add the following two components.
+            Mesh2d(meshes.add(Rectangle::default())),
+            MeshMaterial2d(materials.add(PlayerBaseMaterial {
+                // Generally, only change what's inside the 'lighting' struct and the 'texture' and 'normal' parameters.
+                color: LinearRgba::BLUE,
+                texture: Some(asset_server.load("player/player_albedo.png")),
+                lighting: crate::player_material::Lighting {
+                    // 'ambient_reflection_coefficient' and 'ambient_light_intensity' do the same thing.
+                    // Should be 0 for everything except the player.
+                    // 'diffuse_reflection_coefficient' is how much not-shiny light is reflected back.
+                    // 'shininess' is what it sounds like. Higher number = shinier.
+                    ambient_reflection_coefficient: 0.1,
+                    ambient_light_intensity: 0.1,
+                    diffuse_reflection_coefficient: 1.0,
+                    shininess: 40.0,
+                },
+                lights: lights.lights,
+                normal: Some(asset_server.load("player/player_normal.png")),
+                mesh_rotation: 0.0,
+            })),
+            Transform::from_xyz(0., 0., 0.).with_scale(Vec3::splat(128.)), // Change size of player here: current size: 64. (makes player 64x larger)
+            // you can have a smaller player with 32 and larger player with 128
+            Velocity::new(),
+            FireCooldown(Timer::from_seconds(0.2, TimerMode::Repeating)),
+            Player,
+            Health::new(MAX_HEALTH),
+            //NetControl::new(PlayerType::Local, 0),
+            KinematicCollider {
+                shape: Aabb2d {
+                    min: Vec2 { x: 0., y: 0. },
+                    max: Vec2 { x: 64., y: 64. },
+                },
             },
-            lights: lights.lights,
-            normal: Some(asset_server.load("player/player_normal.png")),
-            mesh_rotation: 0.0,
-        })),
-        Transform::from_xyz(0., 0., 0.).with_scale(Vec3::splat(128.)), // Change size of player here: current size: 64. (makes player 64x larger)
-        // you can have a smaller player with 32 and larger player with 128
-        Velocity::new(),
-        FireCooldown(Timer::from_seconds(0.2, TimerMode::Repeating)),
-        Player,
-        Health::new(MAX_HEALTH),
-        NetControl::new(PlayerControl::Local, 0),
-        KinematicCollider {
-            shape: Aabb2d {
-                min: Vec2 { x: 0., y: 0. },
-                max: Vec2 { x: 64., y: 64. },
-            },
-        },
-    ));
-
-    /*commands.spawn((
-        // For any entities that we want to have lighting,
-        // add the following two components.
-        Mesh2d(meshes.add(Rectangle::default())),
-        MeshMaterial2d(materials.add(PlayerBaseMaterial {
-            // Generally, only change what's inside the 'lighting' struct and the 'texture' and 'normal' parameters.
-            color: LinearRgba::BLUE,
-            texture: Some(asset_server.load("player/player_albedo.png")),
-            lighting: crate::player_material::Lighting {
-                // 'ambient_reflection_coefficient' and 'ambient_light_intensity' do the same thing.
-                // Should be 0 for everything except the player.
-                // 'diffuse_reflection_coefficient' is how much not-shiny light is reflected back.
-                // 'shininess' is what it sounds like. Higher number = shinier.
-                ambient_reflection_coefficient: 0.1,
-                ambient_light_intensity: 0.1,
-                diffuse_reflection_coefficient: 1.0,
-                shininess: 40.0,
-            },
-            lights: lights.lights,
-            normal: Some(asset_server.load("player/player_normal.png")),
-            mesh_rotation: 0.0,
-        })),
-        Transform::from_xyz(0., 0., 0.).with_scale(Vec3::splat(128.)), // Change size of player here: current size: 64. (makes player 64x larger)
-        // you can have a smaller player with 32 and larger player with 128
-        Velocity::new(),
-        FireCooldown(Timer::from_seconds(0.2, TimerMode::Repeating)),
-        Player,
-        Health::new(MAX_HEALTH),
-        NetControl::new(PlayerControl::Network, 1),
-        KinematicCollider {
-            shape: Aabb2d {
-                min: Vec2 { x: 0., y: 0. },
-                max: Vec2 { x: 64., y: 64. },
-            },
-        },
-    ));*/
+        ));
+    }
 }
 
 pub fn player_movement(
@@ -183,11 +127,11 @@ pub fn player_movement(
     input: Res<ButtonInput<KeyCode>>,
     player: Query<(&mut Transform, &mut Velocity, &mut NetControl), With<Player>>,
 ) {
-    //Basic Player
+    //All Player check
     for (mut transform, mut velocity, mut netcontrol) in player {
         let mut dir = Vec2::ZERO;
 
-        if netcontrol.player_type == PlayerControl::Local {
+        if netcontrol.get_type() == PlayerType::Local {
             if input.pressed(KeyCode::KeyA) {
                 dir.x -= 1.;
             }
@@ -204,21 +148,19 @@ pub fn player_movement(
                 dir.y -= 1.;
             }
         } else {
-            let input_str = format!("{:08b}", netcontrol.net_input);
-            //info!(input_str);
-            if input_str.chars().nth(1).unwrap() == '1' {
+            if netcontrol.pressed(KeyCode::KeyA) {
                 dir.x -= 1.;
             }
 
-            if input_str.chars().nth(3).unwrap() == '1' {
+            if netcontrol.pressed(KeyCode::KeyD) {
                 dir.x += 1.;
             }
 
-            if input_str.chars().nth(0).unwrap() == '1' {
+            if netcontrol.pressed(KeyCode::KeyW) {
                 dir.y += 1.;
             }
 
-            if input_str.chars().nth(2).unwrap() == '1' {
+            if netcontrol.pressed(KeyCode::KeyS) {
                 dir.y -= 1.;
             }
         }
@@ -254,7 +196,7 @@ pub fn player_movement(
 }
 
 pub fn player_orientation(
-    mut players: Query<(&mut MeshMaterial2d<PlayerBaseMaterial>, &mut Transform), With<Player>>,
+    mut players: Query<(&mut MeshMaterial2d<PlayerBaseMaterial>, &mut Transform, &mut NetControl), With<Player>>,
     windows: Query<&Window, With<PrimaryWindow>>,
     camera: Query<(&Camera, &GlobalTransform)>,
     asset_server: Res<AssetServer>,
@@ -271,50 +213,26 @@ pub fn player_orientation(
         if let Ok(cursor_world_position) =
             camera.viewport_to_world_2d(camera_transform, cursor_position)
         {
-            for (mut material, mut player_transform) in players.iter_mut() {
-                let player_position = player_transform.translation.truncate();
-                let direction = cursor_world_position - player_position;
+            for (mut material, mut player_transform, mut netcontrol) in players.iter_mut() {
 
-                if direction.length() > 0.0 {
-                    let angle = direction.y.atan2(direction.x);
-                    // let degrees = angle.to_degrees();
+                let mut rounded_rot_z = 0.;
 
-                    // // Convert to 0-360 range and rotate so 0° is bottom
-                    // let normalized_degrees = ((if degrees < 0.0 {
-                    //     degrees + 360.0
-                    // } else {
-                    //     degrees
-                    // }) + 90.0)
-                    //     % 360.0;
+                if netcontrol.player_type == PlayerType::Local {
+                    let player_position = player_transform.translation.truncate();
+                    let direction = cursor_world_position - player_position;
 
-                    let rotation_z = direction.y.atan2(direction.x);
-                    player_transform.rotation = Quat::from_rotation_z(rotation_z - consts::PI / 2.);
+                    if direction.length() > 0.0 {
+                        let rotation_z = direction.y.atan2(direction.x);
+                        //Rounding is needed to prevent precision errors when networking 
+                        rounded_rot_z = (rotation_z * 10.).round()/10.;
+                        netcontrol.set_angle(rounded_rot_z);
+                    }
+                } 
+                else if netcontrol.player_type == PlayerType::Network{
+                    rounded_rot_z = netcontrol.get_angle();
+                } 
 
-                    //     // Map angle ranges to sprite images
-                    //     // Bottom=0°, Right=90°, Top=180°, Left=270°
-                    //     let sprite_path = if normalized_degrees >= 337.5 || normalized_degrees < 22.5 {
-                    //         "player/blueberryman.png" // 0° (bottom)
-                    //     } else if normalized_degrees >= 22.5 && normalized_degrees < 67.5 {
-                    //         "player/blueberryman45.png" // 45° (bottom right)
-                    //     } else if normalized_degrees >= 67.5 && normalized_degrees < 112.5 {
-                    //         "player/blueberryman90.png" // 90° (right)
-                    //     } else if normalized_degrees >= 112.5 && normalized_degrees < 157.5 {
-                    //         "player/blueberryman135.png" // 135° (top right)
-                    //     } else if normalized_degrees >= 157.5 && normalized_degrees < 202.5 {
-                    //         "player/blueberryman180.png" // 180° (top)
-                    //     } else if normalized_degrees >= 202.5 && normalized_degrees < 247.5 {
-                    //         "player/blueberryman-135.png" // 225° (top left)
-                    //     } else if normalized_degrees >= 247.5 && normalized_degrees < 292.5 {
-                    //         "player/blueberryman-90.png" // 270° (left)
-                    //     } else {
-                    //         "player/blueberryman-45.png" // 315° (bottom left)
-                    //     };
-
-                    //     // Get the material handle and update its texture
-                    //     if let Some(material_handle) = materials.get_mut(&material.0) {
-                    //         material_handle.texture = Some(asset_server.load(sprite_path));
-                    //     }
-                }
+                player_transform.rotation = Quat::from_rotation_z(rounded_rot_z - consts::PI / 2.); 
             }
         }
     }
